requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
databricks-sdk==0.102.0
dataclasses-json==0.6.7
docker==7.1.0
etils==1.14.0
fastapi==0.135.3
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
flax==0.12.6
fonttools==4.62.1
fsspec==2026.3.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h5py==3.14.0
huey==2.6.0
humanize==4.15.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jax==0.9.2
jaxlib==0.9.2
jaxtyping==0.3.9
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.5.0
libclang==18.1.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
msgpack==1.1.2
mypy_extensions==1.1.0
namex==0.1.0
numpy==2.4.4
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optax==0.2.8
optree==0.19.0
orbax-checkpoint==0.11.33
orbax-export==0.0.8
packaging==26.0
pandas==2.3.3
pillow==12.2.0
prettytable==3.17.0
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
simplejson==3.20.2
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==1.0.0
tensorflow==2.21.0
tensorstore==0.1.82
termcolor==3.3.0
threadpoolctl==3.6.0
treescope==0.1.10
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
wadler_lindig==0.1.7
wcwidth==0.6.0
Werkzeug==3.1.7
wheel==0.46.3
wrapt==2.1.2
zipp==3.23.0
